+++ lib/alf/relation.rb
@@ -110,11 +110,12 @@ module Alf
 
     def to_hash(from, to=nil)
       if from.is_a?(Hash) and to.nil?
-        raise ArgumentError "Hash of size 1 expected. " unless from.size==1
+        raise ArgumentError "Hash of size 1 expected." unless from.size==1
         to_hash(from.keys.first, from.values.first)
       else
+        from, to = Selection.coerce(from), Selection.coerce(to)
         each.each_with_object({}) do |tuple, hash|
-          key, value = tuple[from], tuple[to]
+          key, value = from.select(tuple), to.select(tuple)
           if hash.has_key?(key) and hash[key] != value
             raise "Key expected for `#{from}`, divergence found on `#{key}`"
           else