+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Make Ordering#compare robust to the presence of `nil`.
 * Ensured Engine::Leaf correctly symbolize tuple keys of its input operand.
 * Selection now also accepts non-array selections from Symbol and String.
+* Relation#to_hash now accepts any Selection-able `from` and `to` args.
 
 # 0.15.0 / 2013-11-01